# docs/swagger.yaml
basePath: /
definitions:
...
  frameworkdto.ErrorDetailDTO:
    description: Error detail structure
    properties:
      code:
        example: ERR_001
        type: string
      details: {}
      message:
        example: An error occurred
        type: string
    type: object
  frameworkdto.ErrorResponseDTO:
    description: Error API response structure
    properties:
      error:
        $ref: '#/definitions/frameworkdto.ErrorDetailDTO'
      success:
        example: false
        type: boolean
    type: object
...
  frameworkdto.GetTenantDTO:
    properties:
      tenant_address:
        type: string
      tenant_email:
        type: string
      tenant_id:
        type: integer
      tenant_name:
        type: string
      tenant_phone:
        type: string
    type: object
...
  frameworkdto.SuccessResponseDTO:
    description: Successful API response structure
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  frameworkdto.UpdateTenantDTO:
    properties:
      tenant_address:
        type: string
      tenant_email:
        type: string
      tenant_name:
        type: string
      tenant_phone:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@geekible.com
    name: API Support
  description: |-
    A comprehensive service framework API with multi-tenant support, user management, authentication, and authorization.

    ## Authentication
    Most endpoints require Bearer token authentication. Include the token in the Authorization header:
    `Authorization: Bearer <your_jwt_token>`

    ## Error Responses
    All error responses follow a standard format with success, error code, and message fields.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Service Framework API
  version: "1.0"
paths:
...
  /tenant/delete:
    delete:
      consumes:
      - application/json
      description: Delete a tenant (requires authentication, authorized roles only)
      produces:
      - application/json
      responses:
        "202":
          description: Tenant deleted successfully
          schema:
            $ref: '#/definitions/frameworkdto.SuccessResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "403":
          description: Not authorized to delete tenant
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Delete tenant
      tags:
      - Tenant
  /tenant/get-all:
    get:
      consumes:
      - application/json
      description: Get list of all tenants (requires authentication, super admin only)
      produces:
      - application/json
      responses:
        "200":
          description: Tenants fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/frameworkdto.SuccessResponseDTO'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/frameworkdto.GetTenantDTO'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "403":
          description: Not authorized to get this resource
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Get all tenants
      tags:
      - Tenant
  /tenant/get-by-id:
    get:
      consumes:
      - application/json
      description: Get tenant details by ID (requires authentication)
      produces:
      - application/json
      responses:
        "200":
          description: Tenant fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/frameworkdto.SuccessResponseDTO'
            - properties:
                data:
                  $ref: '#/definitions/frameworkdto.GetTenantDTO'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "403":
          description: Not authorized to get this resource
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Get tenant by ID
      tags:
      - Tenant
  /tenant/update:
    put:
      consumes:
      - application/json
      description: Update tenant details (requires authentication, authorized roles
        only)
      parameters:
      - description: Tenant update details
        in: body
        name: updateTenantDTO
        required: true
        schema:
          $ref: '#/definitions/frameworkdto.UpdateTenantDTO'
      produces:
      - application/json
      responses:
        "202":
          description: Tenant updated successfully
          schema:
            $ref: '#/definitions/frameworkdto.SuccessResponseDTO'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "403":
          description: Not authorized to update tenant
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/frameworkdto.ErrorResponseDTO'
      security:
      - BearerAuth: []
      summary: Update tenant
      tags:
      - Tenant
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"